Remove entries for item 4180 and 4181
Each of the above item had one file associated with it. Both of the
files were also associated with item 464, as well as a third file -
the three files were the three pages of
"Evr. obshchina v Rossii, pages 1 to 3"
So entries for 4180 and 4181 were removed from the MODS-to_TIFF
mapping.