Remove entries for item 490, 491, 492, 493, 494, and 504.
Each of the above item had one file associated with it. All of the
files are also associated with item 488, as well as some other
files. These are various pages of
"Trial and Conviction of Emma Goldman and Alexander Berkman"
So entries for 490, 491, 492, 493, 494, and 504 were removed
from the MODS-to_TIFF mapping.