Remove entries for item 4179 in MODS-to_TIFF mapping.
The file associated with item 4179 is also associated with
item 752. This file is a picture of the front of a vinyl record
cover. Item 752 also has a second image associated with it,
which is a picture of the back of the record cover.
So the entry for 4179 was removed from the MODS-to_TIFF mapping.